Renamed mechanisms_execute to mech_exec
