+++ .gitignore
@@ -24,3 +24,5 @@ comms_if/target/
 comms_if/.vs/
 
 **/clr_history.txt
+
+**/__pycahce__
